Unify config files.
