+++ app/lib/deepLink.ts
@@ -7,7 +7,7 @@ import { ChapiCredentialRequest } from '../types/chapi';
 import { credentialRequestFromChapiUrl } from './decode';
 import { encodeQueryParams } from './encode';
 import { onShareIntent } from './shareIntent';
-import { LinkConfig } from '../config';
+import { LinkConfig } from '../../config';
 
 /**
  * In order to support OAuth redirects, the Android intent filter was set
