+++ app/lib/registerWallet.ts
@@ -1,5 +1,5 @@
 import { Linking } from 'react-native';
-import { LinkConfig } from '../config';
+import { LinkConfig } from '../../config';
 
 export function registerWallet(): void {
   Linking.canOpenURL(LinkConfig.registerWalletUrl).then(supported => {
